Fix missing biomni.agent.base_agent module (#285)
env_collection.py imported `base_agent` from `biomni.agent.base_agent`,
but that module did not exist — the class was defined inline in
function_generator.py, so importing PaperTaskExtractor raised
ModuleNotFoundError.

Extract `base_agent` into its own module at the path env_collection.py
already expects, and import it from there in function_generator.py.
Both agent subclasses now share one canonical base class.
